--- src/components/Menu/Menu.js
@@ -1,12 +1,0 @@
-import './Menu.css'
-
-function Menu() {
-    return (
-        <main className="main">
-
-            <p>Здесь будет наше меню</p>
-        </main>
-    );
-  }
-  
-  export default Menu;
