+++ src/components/Menu/menu.jsx
@@ -1,0 +1,9 @@
+function Menu() {
+  return (
+    <main>
+      <p>Здесь будет наше меню</p>
+    </main>
+  );
+}
+
+export default Menu;
