+++ src/components/app/app.module.css
@@ -1,0 +1,3 @@
+.app {
+    text-align: center;
+}
