+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from "react-router-dom";
 import './index.css';
-import App from './components/App/App';
+import App from './components/App/app.jsx';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
